// src/Features/orders/types/orderstypes.ts
export type OrdersResponse = Order[]

export interface Order {
  _id: string
  id: number
  shippingAddress: ShippingAddress
  taxPrice: number
  shippingPrice: number
  totalOrderPrice: number
  paymentMethodType: "cash" | "card"
  isPaid: boolean
  isDelivered: boolean
  createdAt: string
  updatedAt: string
  paidAt?: string
  __v: number
  user: User
  cartItems: CartItem[]
}

export interface ShippingAddress {
  details: string
  phone: string
  city: string
}

export interface User {
  _id: string
  name: string
  email: string
  phone: string
}

export interface CartItem {
  _id: string
  count: number
  price: number
  product: Product
}

export interface Product {
  _id: string
  id: string
  title: string
  imageCover: string
  ratingsAverage: number
  ratingsQuantity: number
  category: Category
  brand: Brand
  subcategory: SubCategory[]
}

export interface Category {
  _id: string
  name: string
  slug: string
  image: string
}

export interface Brand {
  _id: string
  name: string
  slug: string
  image: string
}

export interface SubCategory {
  _id: string
  name: string
  slug: string
  category: string
}